refactor explain_insert into plan/insert.rs explain
